fix: Fix blur event on `Composite` for React 17
