+++ package.json
@@ -5,6 +5,11 @@
       "packages/*"
     ]
   },
+  "resolutions": {
+    "**/react": "17.0.0-rc.0",
+    "**/react-dom": "17.0.0-rc.0",
+    "**/react-test-renderer": "16.13.1"
+  },
   "scripts": {
     "test": "jest",
     "test:all": "yarn lint && yarn build && yarn type-check && yarn test && yarn website:build && yarn clean && yarn website:serve",
