+++ packages/reakit-playground/src/__deps/reakit-utils.ts
@@ -6,8 +6,8 @@ export default {
   "reakit-utils/closest": require("reakit-utils/closest"),
   "reakit-utils/contains": require("reakit-utils/contains"),
   "reakit-utils/createEvent": require("reakit-utils/createEvent"),
-  "reakit-utils/createKeyboardEvent": require("reakit-utils/createKeyboardEvent"),
   "reakit-utils/createOnKeyDown": require("reakit-utils/createOnKeyDown"),
+  "reakit-utils/fireBlurEvent": require("reakit-utils/fireBlurEvent"),
   "reakit-utils/fireEvent": require("reakit-utils/fireEvent"),
   "reakit-utils/fireKeyboardEvent": require("reakit-utils/fireKeyboardEvent"),
   "reakit-utils/flatten": require("reakit-utils/flatten"),
