+++ packages/reakit-utils/src/fireKeyboardEvent.ts
@@ -1,4 +1,30 @@
-import { createKeyboardEvent } from "./createKeyboardEvent";
+import { getDocument } from "./getDocument";
+import { getDefaultView } from "./getDefaultView";
+
+function createKeyboardEvent(
+  element: HTMLElement,
+  type: string,
+  eventInit: KeyboardEventInit = {}
+): KeyboardEvent {
+  if (typeof KeyboardEvent === "function") {
+    return new KeyboardEvent(type, eventInit);
+  }
+  // IE 11 doesn't support Event constructors
+  const event = getDocument(element).createEvent("KeyboardEvent");
+  (event as any).initKeyboardEvent(
+    type,
+    eventInit.bubbles,
+    eventInit.cancelable,
+    getDefaultView(element),
+    eventInit.key,
+    eventInit.location,
+    eventInit.ctrlKey,
+    eventInit.altKey,
+    eventInit.shiftKey,
+    eventInit.metaKey
+  );
+  return event;
+}
 
 /**
  * Creates and dispatches `KeyboardEvent` in a way that also works on IE 11.
