+++ packages/reakit-utils/src/index.ts
@@ -2,9 +2,9 @@ export * from "./applyState";
 export * from "./closest";
 export * from "./contains";
 export * from "./createEvent";
-export * from "./createKeyboardEvent";
 export * from "./createOnKeyDown";
 export * from "./fireEvent";
+export * from "./fireBlurEvent";
 export * from "./fireKeyboardEvent";
 export * from "./flatten";
 export * from "./getActiveElement";
